Split file matching code out of Stack and into a new Matcher class
